# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/delete": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "description": "用户id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.SuccApiExamDel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.ErrApiExamDel"
                        }
                    }
                }
            }
        },
        "/api/delqiniufile": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "删除七牛文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件名称",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.SuccApiDelQiNiuFile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.ErrApiDelQiNiuFile"
                        }
                    }
                }
            }
        },
        "/api/query": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.SuccApiExamQuery"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.ErrApiExamQuery"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "conf.ErrApiDelQiNiuFile": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "删除失败"
                }
            }
        },
        "conf.ErrApiExamDel": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "删除失败"
                }
            }
        },
        "conf.ErrApiExamQuery": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "获取列表出错"
                }
            }
        },
        "conf.SuccApiDelQiNiuFile": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "string",
                    "example": "xxx.png"
                },
                "msg": {
                    "type": "string",
                    "example": "删除成功"
                }
            }
        },
        "conf.SuccApiExamDel": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "msg": {
                    "type": "string",
                    "example": "删除成功"
                }
            }
        },
        "conf.SuccApiExamQuery": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/conf.SuccApiExamQueryData"
                },
                "msg": {
                    "type": "string",
                    "example": "获取成功"
                }
            }
        },
        "conf.SuccApiExamQueryData": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/conf.User"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "conf.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "地址示例"
                },
                "age": {
                    "type": "integer",
                    "example": 20
                },
                "createTime": {
                    "type": "string",
                    "example": "2022-12-12 00:00:00"
                },
                "id": {
                    "type": "integer"
                },
                "sex": {
                    "type": "string",
                    "example": "男"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1661233245
                },
                "username": {
                    "type": "string",
                    "example": "用户名"
                }
            }
        }
    }
}